몬스터 스케일, 스텟 조정 등

+++ Assets/02_Scripts/ImMinhyuk/Monster/MonsterSubClass/IDamagable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IDamagable
 {
-    void Damaged(int damage);
+    void Damaged(float damage);
 }
